backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
supabase
postgrest
httpx
Pillow  # (for image sanity checks if needed)
